# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "docket",
    "description": "Extract, classify and validate structured data from business documents.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/process": {
      "post": {
        "summary": "Process Document",
        "operationId": "process_document_process_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_process_document_process_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_jobs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_jobs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Job Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review-queue": {
      "get": {
        "summary": "Get Review Queue",
        "operationId": "get_review_queue_review_queue_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Review Queue Review Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review-queue/{document_id}": {
      "get": {
        "summary": "Get Review",
        "operationId": "get_review_review_queue__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Review Review Queue  Document Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Review",
        "operationId": "update_review_review_queue__document_id__patch",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Review Review Queue  Document Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review-queue/{document_id}/original": {
      "get": {
        "summary": "Get Review Original",
        "operationId": "get_review_original_review_queue__document_id__original_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_job_jobs_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_job_jobs_post"
      },
      "Body_process_document_process_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_process_document_process_post"
      },
      "ClassificationResult": {
        "properties": {
          "doc_type": {
            "$ref": "#/components/schemas/DocType"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Scores"
          }
        },
        "type": "object",
        "required": [
          "doc_type",
          "confidence",
          "method"
        ],
        "title": "ClassificationResult"
      },
      "DocType": {
        "type": "string",
        "enum": [
          "invoice",
          "receipt",
          "contract",
          "purchase_order",
          "bank_statement",
          "acceptance_act",
          "waybill",
          "boarding_pass",
          "unknown"
        ],
        "title": "DocType"
      },
      "DocumentForensicReport": {
        "properties": {
          "has_signatures": {
            "type": "boolean",
            "title": "Has Signatures",
            "description": "Whether at least one handwritten signature is present.",
            "default": false
          },
          "has_stamps": {
            "type": "boolean",
            "title": "Has Stamps",
            "description": "Whether at least one seal or official stamp is present.",
            "default": false
          },
          "is_executed": {
            "type": "boolean",
            "title": "Is Executed",
            "description": "Whether the document appears legally executed (signed/stamped).",
            "default": false
          },
          "is_empty_template": {
            "type": "boolean",
            "title": "Is Empty Template",
            "description": "Whether the document appears to be an unexecuted blank template.",
            "default": false
          },
          "stamps": {
            "items": {
              "$ref": "#/components/schemas/StampDetection"
            },
            "type": "array",
            "title": "Stamps"
          },
          "signatures": {
            "items": {
              "$ref": "#/components/schemas/SignatureDetection"
            },
            "type": "array",
            "title": "Signatures"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/HandwrittenAnnotation"
            },
            "type": "array",
            "title": "Annotations"
          },
          "alterations_detected": {
            "type": "boolean",
            "title": "Alterations Detected",
            "description": "Whether handwritten overrides or crossed-out amounts were detected.",
            "default": false
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Flags",
            "description": "Forensic risk flags (e.g. UNEXECUTED_TEMPLATE, MISSING_STAMP, HANDWRITTEN_ALTERATION)."
          }
        },
        "type": "object",
        "title": "DocumentForensicReport",
        "description": "Forensic report covering physical execution, stamps, signatures, and alterations."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineResult": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult"
          },
          "extracted": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted"
          },
          "field_sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceLocation"
            },
            "type": "object",
            "title": "Field Sources",
            "description": "Where each extracted field was read from — kept beside the data, not inside it."
          },
          "extract_attempts": {
            "type": "integer",
            "title": "Extract Attempts"
          },
          "validation_issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Validation Issues"
          },
          "ocr_method": {
            "type": "string",
            "title": "Ocr Method"
          },
          "raw_text_chars": {
            "type": "integer",
            "title": "Raw Text Chars"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "unknown"
          },
          "escalated_to_vlm": {
            "type": "boolean",
            "title": "Escalated To Vlm",
            "default": false
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review",
            "default": false
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls",
            "default": 0
          },
          "llm_estimated_tokens": {
            "type": "integer",
            "title": "Llm Estimated Tokens",
            "default": 0
          },
          "pages_total": {
            "type": "integer",
            "title": "Pages Total",
            "default": 1
          },
          "pages_processed": {
            "type": "integer",
            "title": "Pages Processed",
            "default": 1
          },
          "complete": {
            "type": "boolean",
            "title": "Complete",
            "default": true
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "page_methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Page Methods"
          },
          "forensic_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentForensicReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "source",
          "classification",
          "extracted",
          "extract_attempts",
          "ocr_method",
          "raw_text_chars"
        ],
        "title": "PipelineResult"
      },
      "ReviewUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "corrections": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrections"
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "reviewer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ReviewUpdate"
      },
      "SignatureDetection": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "1-indexed page number where signature was found."
          },
          "box": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Box",
            "description": "Normalized bounding box (ymin, xmin, ymax, xmax) from 0.0 to 1.0."
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Signature detection confidence score.",
            "default": 1.0
          },
          "signatory_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signatory Label",
            "description": "Contextual role (e.g. Customer, Contractor, Director) if labeled."
          }
        },
        "type": "object",
        "required": [
          "page",
          "box"
        ],
        "title": "SignatureDetection",
        "description": "Detection of a handwritten signature, mark, or cursive endorsement."
      },
      "SourceLocation": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "quote": {
            "type": "string",
            "minLength": 1,
            "title": "Quote",
            "description": "Exact text copied from the source page."
          }
        },
        "type": "object",
        "required": [
          "page",
          "quote"
        ],
        "title": "SourceLocation",
        "description": "Auditable link from a field to the original page region."
      },
      "StampDetection": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "1-indexed page number where stamp was found."
          },
          "box": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Box",
            "description": "Normalized bounding box (ymin, xmin, ymax, xmax) from 0.0 to 1.0."
          },
          "color": {
            "type": "string",
            "title": "Color",
            "description": "Dominant ink color (e.g. blue, violet, red, black).",
            "default": "blue"
          },
          "shape": {
            "type": "string",
            "title": "Shape",
            "description": "Geometry of the stamp (circular, oval, rectangular, irregular).",
            "default": "circular"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Detection confidence score.",
            "default": 1.0
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Decoded text or organization name from stamp, if legible."
          }
        },
        "type": "object",
        "required": [
          "page",
          "box"
        ],
        "title": "StampDetection",
        "description": "Detection of a physical seal, official rubber stamp, or organization stamp."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationIssue": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "error"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ValidationIssue"
      }
    }
  }
}